extract occurs check into its own module
